requirements.txt
streamlit
ultralytics
opencv-python
pillow
pymongo
numpy
